+++ storm-core/src/clj/org/apache/storm/ui/core.clj
@@ -164,11 +164,6 @@
 (defn supervisor-log-link [host]
   (UIHelpers/urlFormat "http://%s:%s/daemonlog?file=supervisor.log" (to-array [host (*STORM-CONF* LOGVIEWER-PORT)])))
 
-(defn get-error-time
-  [error]
-  (if error
-    (Time/deltaSecs (.get_error_time_secs ^ErrorInfo error))))
-
 (defn get-error-data
   [error]
   (if error
@@ -190,8 +185,7 @@
 (defn get-error-time
   [error]
   (if error
-    (.get_error_time_secs ^ErrorInfo error)
-    ""))
+    (.get_error_time_secs ^ErrorInfo error)))
 
 (defn worker-dump-link [host port topology-id]
   (UIHelpers/urlFormat "http://%s:%s/dumps/%s/%s"
@@ -540,7 +534,7 @@
      "errorTime" (get-error-time error-info)
      "errorHost" host
      "errorPort" port
-     "errorLapsedSecs" (get-error-time error-info)
+     "errorLapsedSecs" (if-let [t (get-error-time error-info)] (Time/deltaSecs t))
      "errorWorkerLogLink" (worker-log-link host port topo-id secure?)}))
 
 (defn- common-agg-stats-json
@@ -666,14 +660,14 @@
                     reverse)]
     {"componentErrors"
      (for [^ErrorInfo e errors]
-       {"errorTime" (* 1000 (long (.get_error_time_secs e)))
+       {"errorTime" (get-error-time e)
         "errorHost" (.get_host e)
         "errorPort"  (.get_port e)
         "errorWorkerLogLink"  (worker-log-link (.get_host e)
                                                (.get_port e)
                                                topology-id
                                                secure?)
-        "errorLapsedSecs" (get-error-time e)
+        "errorLapsedSecs" (if-let [t (get-error-time e)] (Time/deltaSecs t))
         "error" (.get_error e)})}))
 
 (defmulti unpack-comp-agg-stat