+++ CHANGELOG.md
@@ -39,6 +39,7 @@
  * STORM-1521: When using Kerberos login from keytab with multiple bolts/executors ticket is not renewed in hbase bolt.
 
 ## 1.0.0
+ * STORM-1488: UI Topology Page component last error timestamp is from 1970
  * STORM-1552: Fix topology event sampling log dir
  * STORM-1542: Remove profile action retry in case of non-zero exit code
  * STORM-1540: Fix Debug/Sampling for Trident